+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.obstaclerule.DirectionAttackPiece;
+import chess.domain.piece.attacker.DirectionAttackPiece;
 
 public class Bishop extends DirectionAttackPiece {
     public Bishop(final Color color) {
