+++ src/main/java/chess/domain/piece/Empty.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.obstaclerule.NotAttackPiece;
+import chess.domain.piece.attacker.NotAttackPiece;
 import chess.domain.position.Position;
 import java.util.Map;
