+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.obstaclerule.DirectionAttackPiece;
+import chess.domain.piece.attacker.DirectionAttackPiece;
 
 public class Queen extends DirectionAttackPiece {
     public Queen(final Color color) {
