+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.obstaclerule.DirectionAttackPiece;
+import chess.domain.piece.attacker.DirectionAttackPiece;
 
 public class Rook extends DirectionAttackPiece {
     public Rook(final Color color) {
